Move type helper functions to extra class
