+++ src/rmet/batch_evaluator.py
@@ -12,6 +12,9 @@ from .metrics import (
     supported_user_accuracy_metrics,
     supported_global_metrics,
     supported_user_beyond_accuracy_metrics,
+)
+
+from .type_helpers import (
     _std,
 )
